I did lots of things. Interface mostly.
It generates the files now in PDF by default. There is an option to
change the output format. Every option now is seeable with the help
option. Help is also printed on some error of parsing the command line
arguments. There is an output option to redirect the output into a file.
Added automatic dependency management with gcc.